+++ src/main/java/application_state/FakeEventBus.java
@@ -1,0 +1,6 @@
+package application_state;
+
+public class FakeEventBus {
+
+
+}
