+++ src/main/java/home/DirectionsController.java
@@ -9,10 +9,12 @@ import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
+import javafx.geometry.Pos;
 import javafx.scene.control.Label;
 import javafx.scene.paint.Color;
 import application_state.Event;
 import application_state.EventBusFactory;
+import map.MapController;
 import map.Node;
 import service.TextingService;
 
@@ -33,7 +35,7 @@ public class DirectionsController {
     private JFXButton textingButton;
 
     //text message global variable
-    private String units = "Ft";    // Feet or meters conversion
+    private String units = "feet";    // Feet or meters conversion
     private HashMap<String, Integer> floors = new HashMap<String, Integer>();
     private ArrayList<Node> path;
 
@@ -41,6 +43,9 @@ public class DirectionsController {
     @FXML
     void initialize() {
         eventBus.register(this);
+        System.out.println("initialize");
+        path = event.getPath();
+        printDirections(makeDirections(path));
     }
 
     @FXML
@@ -50,26 +55,26 @@ public class DirectionsController {
     }
 
 
-    @Subscribe
-    void eventListener(Event newevent) {
-        Platform.runLater(new Runnable() {
-            @Override
-            public void run() {
-                switch (event.getEventName()) {
-                    case "printText":
-//                        Thread.sleep(100);
-                        path = event.getPath();
-                        printDirections(makeDirections(path));
-                        break;
-                    default:
-                        break;
-                }
-            }
-        });
-    }
+//    @Subscribe todo
+//    void eventListener(Event newevent) {
+//        Platform.runLater(new Runnable() {
+//            @Override
+//            public void run() {
+//                switch (event.getEventName()) {
+//                    case "printText":
+////                        Thread.sleep(100);
+//                        path = event.getPath();
+//                        printDirections(makeDirections(path));
+//                        break;
+//                    default:
+//                        break;
+//                }
+//            }
+//        });
+//    }
 
 
-
+    // TODO: why does 1st instruction turn into ... if not scrollable?
     /**
      * Create textual instructions for the given path.
      * @param path the list of nodes in the path
@@ -93,16 +98,16 @@ public class DirectionsController {
         ArrayList<String> directions = new ArrayList<>();    // Collection of instructions
         directions.add("\nStart at " + path.get(0).getLongName() + ".\n");    // First instruction
 
-        // Make the next instruction cardinal, or up/down if it is a floor connector
+        // Make the first instruction cardinal, or up/down if it is a floor connector
         String oldFloor = path.get(0).getFloor();
         String newFloor = path.get(1).getFloor();
         if (!floors.get(oldFloor).equals(floors.get(newFloor))) {
             directions.add(upDownConverter(oldFloor, newFloor, path.get(0).getNodeType()));
         }
-        else if ((path.size() == 2 && path.get(1).getNodeType().equals("ELEV")) || (path.size() > 2 && path.get(2).getNodeType().equals("ELEV"))) {
+        else if ((path.size() == 2 && path.get(1).getNodeType().equals("ELEV")) || (path.size() > 2 && path.get(1).getNodeType().equals("ELEV") && path.get(2).getNodeType().equals("ELEV"))) {
             directions.add("I");
         }
-        else if ((path.size() == 2 && path.get(1).getNodeType().equals("STAI")) || (path.size() > 2 && path.get(2).getNodeType().equals("STAI"))) {
+        else if ((path.size() == 2 && path.get(1).getNodeType().equals("STAI")) || (path.size() > 2 && path.get(1).getNodeType().equals("STAI") && path.get(2).getNodeType().equals("STAI"))) {
             directions.add("J");
         }
         else {
@@ -135,6 +140,7 @@ public class DirectionsController {
             }
         }
 
+        System.out.println("before simplifying: " + directions);
         // Simplify directions that continue approximately straight from each other
         for (int i = 1; i < directions.size(); i++) {
             String currDir = directions.get(i);
@@ -144,10 +150,10 @@ public class DirectionsController {
             String newDir = "";
             boolean changed = false;
             if (currOne.equals("A") && !"IJ".contains(prevOne)) {
-                int prevDist = Integer.parseInt(prevDir.substring(1));
-                int currDist = Integer.parseInt(currDir.substring(1));
-                int totalDist = prevDist + currDist;    // Combine the distance of this direction with the previous one
-                newDir = prevOne + totalDist;
+                int prevDist = Integer.parseInt(prevDir.substring(1,6));
+                int currDist = Integer.parseInt(currDir.substring(1,6));
+                double totalDist = prevDist + currDist;    // Combine the distance of this direction with the previous one
+                newDir = prevOne + padWithZeros(totalDist) + currDir.substring(6);
                 changed = true;
             }
             else if ("NOPQ".contains(currOne) && currOne.equals(prevOne)) {    // If the current direction contains straight, get the distance substring
@@ -164,6 +170,7 @@ public class DirectionsController {
 
         // Add the final direction
         directions.add("You have arrived at " + path.get(path.size() - 1).getLongName() + ".");
+        System.out.println(directions); // TODO cut
         return directions;
     }
 
@@ -223,34 +230,35 @@ public class DirectionsController {
         Label first = new Label(ds.get(0));
         first.setWrapText(true);
         first.setTextFill(Color.WHITE);
+        //first.setPrefWidth(450);
         labels.add(first);
 
         for (int i = 1; i < ds.size() - 1; i++) {
             String direct = ds.get(i);
             switch(direct.substring(0,1)) {
                 case "A":
-                    direct = "Walk straight for " + direct.substring(1) + " " + units + ".\n";
+                    direct = "Walk straight for " + Integer.parseInt(direct.substring(1,6)) + " " + units + direct.substring(6) + ".\n";
                     break;
                 case "B":
-                    direct = "Turn left and walk for " + direct.substring(1) + " " + units + ".\n";
+                    direct = "Turn left and walk for " + Integer.parseInt(direct.substring(1,6)) + " " + units  + direct.substring(6) + ".\n";
                     break;
                 case "C":
-                    direct = "Turn slightly left and walk for " + direct.substring(1) + " " + units + ".\n";
+                    direct = "Turn slightly left and walk for " + Integer.parseInt(direct.substring(1,6)) + " " + units  + direct.substring(6) + ".\n";
                     break;
                 case "D":
-                    direct = "Turn sharply left and walk for " + direct.substring(1) + " " + units + ".\n";
+                    direct = "Turn sharply left and walk for " + Integer.parseInt(direct.substring(1,6)) + " " + units  + direct.substring(6) + ".\n";
                     break;
                 case "E":
-                    direct = "Turn right and walk for " + direct.substring(1) + " " + units + ".\n";
+                    direct = "Turn right and walk for " + Integer.parseInt(direct.substring(1,6)) + " " + units  + direct.substring(6) + ".\n";
                     break;
                 case "F":
-                    direct = "Turn slightly right and walk for " + direct.substring(1) + " " + units + ".\n";
+                    direct = "Turn slightly right and walk for " + Integer.parseInt(direct.substring(1,6)) + " " + units  + direct.substring(6) + ".\n";
                     break;
                 case "G":
-                    direct = "Turn sharply right and walk for " + direct.substring(1) + " " + units + ".\n";
+                    direct = "Turn sharply right and walk for " + Integer.parseInt(direct.substring(1,6)) + " " + units  + direct.substring(6) + ".\n";
                     break;
                 case "H":
-                    direct = "Turn around and walk for " + direct.substring(1) + " " + units + ".\n";
+                    direct = "Turn around and walk for " + Integer.parseInt(direct.substring(1,6)) + " " + units  + direct.substring(6) + ".\n";
                     break;
                 case "I":
                     direct = "Walk to the elevator.\n";
@@ -271,28 +279,28 @@ public class DirectionsController {
                     direct = "Take the stairs down from floor " + backToFloors.get(direct.substring(1,2)) + " to floor " + backToFloors.get(direct.substring(2,3)) + ".\n";
                     break;
                 case "S":
-                    direct = "Walk north for " + direct.substring(1) + " " + units + ".\n";
+                    direct = "Walk north for " + Integer.parseInt(direct.substring(1,6)) + " " + units  + direct.substring(6) + ".\n";
                     break;
                 case "T":
-                    direct = "Walk north west for " + direct.substring(1) + " " + units + ".\n";
+                    direct = "Walk north west for " + Integer.parseInt(direct.substring(1,6)) + " " + units  + direct.substring(6) + ".\n";
                     break;
                 case "U":
-                    direct = "Walk west for " + direct.substring(1) + " " + units + ".\n";
+                    direct = "Walk west for " + Integer.parseInt(direct.substring(1,6)) + " " + units  + direct.substring(6) + ".\n";
                     break;
                 case "V":
-                    direct = "Walk south west for " + direct.substring(1) + " " + units + ".\n";
+                    direct = "Walk south west for " + Integer.parseInt(direct.substring(1,6)) + " " + units  + direct.substring(6) + ".\n";
                     break;
                 case "W":
-                    direct = "Walk south for " + direct.substring(1) + " " + units + ".\n";
+                    direct = "Walk south for " + Integer.parseInt(direct.substring(1,6)) + " " + units  + direct.substring(6) + ".\n";
                     break;
                 case "X":
-                    direct = "Walk south east for " + direct.substring(1) + " " + units + ".\n";
+                    direct = "Walk south east for " + Integer.parseInt(direct.substring(1,6)) + " " + units  + direct.substring(6) + ".\n";
                     break;
                 case "Y":
-                    direct = "Walk east for " + direct.substring(1) + " " + units + ".\n";
+                    direct = "Walk east for " + Integer.parseInt(direct.substring(1,6)) + " " + units  + direct.substring(6) + ".\n";
                     break;
                 case "Z":
-                    direct = "Walk north east for " + direct.substring(1) + " " + units + ".\n";
+                    direct = "Walk north east for " + Integer.parseInt(direct.substring(1,6)) + " " + units  + direct.substring(6) + ".\n";
                     break;
                 default:
                     direct = "Houston we have a problem";
@@ -301,6 +309,7 @@ public class DirectionsController {
 
             Label l = new Label(direct);
             l.setWrapText(true);
+           // l.setPrefWidth(450); todo: get text to wrap
             l.setTextFill(Color.WHITE);
             labels.add(l);
             directions.add(direct);
@@ -309,6 +318,7 @@ public class DirectionsController {
 
         Label last = new Label(ds.get(ds.size() - 1));
         last.setWrapText(true);
+       // last.setPrefWidth(450);
         last.setTextFill(Color.WHITE);
         labels.add(last);
 
@@ -322,9 +332,9 @@ public class DirectionsController {
             buf.append(directions.get(i));
         }
         String total = buf.toString();
+        System.out.println(total);
         return total;
     }
-
 
     /**
      * Convert this direction to a cardinal direction
@@ -404,7 +414,7 @@ public class DirectionsController {
 
         // Distance in feet based on measurements from the map: 260 pixels per 85 feet
         double distance;
-        if (units.equals("Ft")) {
+        if (units.equals("feet")) {
             distance = lengthNew /260 * 85;    // Pixels to feet
         }
         else {
@@ -476,10 +486,71 @@ public class DirectionsController {
             }
         }
 
+        // TODO: landmarks
+        String landmark = "";
+        ArrayList<Node> closeNodes = MapController.getNodesConnectedTo(next);
+        //  ArrayList<Node> closeNodes = DatabaseService.getDatabaseService().getNodesConnectedTo(next);
+        for (Node n: closeNodes) {
+            if (!landmark.contains("towards") && n.getNodeType().equals("HALL")) {
+                landmark = " down the hall";
+            }
+            if (!n.getNodeType().equals("HALL") && !n.getNodeType().equals("ELEV") && !n.getNodeType().equals("STAI")) {
+                landmark = " towards " + n.getLongName();
+            }
+        }
+        if (!next.getNodeType().equals("HALL") && !next.getNodeType().equals("ELEV") && !next.getNodeType().equals("STAI")) {
+            landmark = " towards " + next.getLongName();
+        }
+
         // Create and return the direction
-        String direction = String.format(turn +  "%.0f", distance);
+        String distPadded = padWithZeros(distance);    // Pad direction with zeros so always same lengtb
+        String direction = turn + distPadded + landmark;
         return direction;
     }
+
+    /**
+     * Pad a distance with zeros so all distances are the same length
+     * @param distance the distance to pad with zeros
+     * @return the padded distance as a string
+     */
+    private String padWithZeros(double distance) {
+        String orig = String.format("%.0f", distance);
+        for (int i = 0; i < 6; i++) {    // Assume max distance is less than 99999 feet
+            if (orig.length() < i) {    // Pad with zeroes while string length is less than 6
+                orig = "0" + orig;
+            }
+        }
+        return orig;
+    }
+
+    /**
+     * On click, show the directions. On second click, hide them again.
+     */
+//    @FXML
+//    public void showDirections() {
+//        directionsView.setVisible(!directionsView.isVisible());
+//        if (showDirectionsBtn.getText().contains("Show")) {
+//            showDirectionsBtn.setText("Close Textual Directions");
+//            directionsView.toFront();
+//        }
+//        else {
+//            showDirectionsBtn.setText("Show Textual Directions");
+//            showDirVbox.toFront();
+//        }
+//        if (showDirVbox.getAlignment().equals(Pos.BOTTOM_RIGHT)) {
+//            showDirVbox.setAlignment(Pos.TOP_RIGHT);
+//        }
+//        else {
+//            showDirVbox.setAlignment(Pos.BOTTOM_RIGHT);
+//        }
+//    }   todo hmmm still needed or not?
+
+//    private void hideDirections() {
+//        showDirectionsBtn.setText("Show Textual Directions");
+//        showDirVbox.toFront();
+//        showDirVbox.setAlignment(Pos.BOTTOM_RIGHT);
+//        directionsView.setVisible(false);
+//    }
 
     /**
      * Get the current units
@@ -489,19 +560,25 @@ public class DirectionsController {
         return units;
     }
 
+    // TODO: add all changes to directions controller... blech. also cut print statements.
+
     /**
      * Set the current units as feet or meters
      */
     public void setUnits() {
-        if (unitSwitch_btn.getText().equals("M")) {
-            units = "M";
-            unitSwitch_btn.setText("Ft");
+        if (unitSwitch_btn.getText().equals("Meters")) {
+            units = "meters";
+            unitSwitch_btn.setText("Feet");
         }
         else {
-            units = "Ft";
-            unitSwitch_btn.setText("M");
+            units = "feet";
+            unitSwitch_btn.setText("Meters");
         }
         printDirections(makeDirections(path));
+    }
+
+    public void setPath(ArrayList<Node> thePath) {
+        this.path = thePath;
     }
 
     /**
